Enum reflection (#989)
* Enum support in reflection

-Cleanup naming in code-gen
-Add a enum type parser
-Add enum codegen
-Add a enum type to gRPC messages to the Editor

* Regen sample-data + snapshot with new Enum type

+++ crates/lgn-asset-registry/src/asset_to_ecs.rs
@@ -99,7 +99,7 @@ impl AssetToECS for runtime_data::Entity {
             } else if let Some(static_mesh) = component.downcast_ref::<runtime_data::StaticMesh>() {
                 entity.insert(StaticMesh::from_default_meshes(
                     default_meshes,
-                    static_mesh.mesh_id,
+                    static_mesh.mesh_id as usize,
                     static_mesh.color,
                     DefaultMaterialType::Default,
                 ));
